#15 Staggered view seems good to go now. but need some tweeks

+++ WinstaCore/ApplicationSettingsManager.cs
@@ -236,7 +236,7 @@ public class ApplicationSettingsManager
 
     public ApplicationLockMode SetAppLockMode(ApplicationLockMode lockMode = ApplicationLockMode.WindowsHello)
     {
-        LocalSettings.Values[AppLockMode] = lockMode;
+        LocalSettings.Values[AppLockMode] = (int)lockMode;
         return lockMode;
     }
